chore(IAM-173): separating out some timezone code for code cleanliness
